refactor: Incorporate feedback
