+++ src/main/java/com/mide/gangsaeng/board/BoardRdbRepositoryImpl.java
@@ -6,11 +6,11 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public class BoardRepositoryImpl implements BoardRepository {
+public class BoardRdbRepositoryImpl implements BoardRepository {
 
     private final BoardMapper mapper;
 
-    public BoardRepositoryImpl(BoardMapper mapper) {this.mapper = mapper;}
+    public BoardRdbRepositoryImpl(BoardMapper mapper) {this.mapper = mapper;}
 
     public void write(BoardRequest request) {
         mapper.write(request);
