--- src/main/java/com/mide/gangsaeng/board/BoardReadStorage.java
@@ -1,5 +1,0 @@
-package com.mide.gangsaeng.board;
-
-public interface BoardReadStorage {
-    Board read(long id);
-}
